Bug 1237023: Update of cherry-pick of VP9: add some base files, compiles-and-works r=pkerr
UltraBlame original commit: 4bd447a3d7da016d2a3bb0f5de0511cfdf40113e

+++ media/webrtc/trunk/webrtc/modules/video_capture/linux/device_info_linux.cc
@@ -164,7 +164,7 @@ int32_t DeviceInfoLinux::GetDeviceName(
     } else {
         
         if (snprintf(deviceUniqueIdUTF8, deviceUniqueIdUTF8Length, "fake_%u", device_index) >=
-            deviceUniqueIdUTF8Length)
+            (int) deviceUniqueIdUTF8Length)
         {
             WEBRTC_TRACE(webrtc::kTraceError, webrtc::kTraceVideoCapture, _id,
                        "buffer passed is too small");
